requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pandas==2.1.4
numpy==1.26.2
python-multipart==0.0.6
requests==2.31.0
python-dateutil==2.8.2
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==11.0.3
click==8.1.7
